# Conflicts:
#	src/main/resources/templates/wishlist/viewWishlist.html

+++ src/main/java/com/example/wishlist/repository/WishRepository.java
@@ -9,5 +9,5 @@ public interface WishRepository {
     Wish getWish(long id);
     boolean deleteWish(long id);
     boolean editWish(Wish wish);
-    boolean deleteSelectedWishes(List<Integer> wishIdList);
+    boolean deleteSelectedWishes(List<Long> wishIdList);
 }
